+++ src/core/parsers/composite.rs
@@ -1,5 +1,6 @@
 use crate::core::{
     error::{Error, Result},
+    fs,
     parsers::{RawItem, RawKind},
     profile::{Profile, ProfileType, composite::Composite},
 };
@@ -19,11 +20,24 @@ impl CompositeParser {
             match line.kind {
                 // composite profile has NO options lines
                 RawKind::Option => {
-                    return Err(Error::InvalidOptionLine(profile, line.line, line.content));
+                    return Err(Error::InvalidOptionLine(
+                        profile,
+                        line.line,
+                        line.content,
+                        "".into(),
+                    ));
                 }
 
                 // normal data lines, aka profile names here
                 RawKind::Data => {
+                    if fs::check_has_parent_dirs(&line.content) {
+                        return Err(Error::InvalidDataLine(
+                            profile,
+                            line.line,
+                            line.content,
+                            "profile name cannot be paths with parent directories".into(),
+                        ));
+                    }
                     entries.push(line.content);
                 }
             }