+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented here
 ## v0.8.1
 
 - updated colors for rmhome and rmbackup paths
-- [WIP] parser does not allow .. paths anymore
+- parser does not allow .. paths anymore
 - [WIP] binary now checks no symlink that links to outside the repo exist before running 
 
 ## v0.8.0